Refactored signature verification to composable_support
