+++ frame/airdrop/src/lib.rs
@@ -28,7 +28,7 @@ pub mod pallet {
 			},
 		},
 		math::safe::{SafeAdd, SafeSub},
-		types::{CosmosAddress, CosmosEcdsaSignature, EcdsaSignature, EthereumAddress},
+		signature_verification,
 	};
 	use composable_traits::airdrop::Airdropper;
 	use frame_support::{
@@ -41,15 +41,13 @@ pub mod pallet {
 		transactional, Blake2_128Concat, PalletId, Parameter,
 	};
 	use frame_system::pallet_prelude::*;
-	use p256::ecdsa::{signature::Verifier, Signature, VerifyingKey};
 	use scale_info::TypeInfo;
-	use sp_io::hashing::{keccak_256, sha2_256};
 	use sp_runtime::{
 		traits::{
 			AccountIdConversion, AtLeast32Bit, AtLeast32BitUnsigned, CheckedAdd, CheckedMul,
-			CheckedSub, Convert, One, Saturating, Verify, Zero,
+			CheckedSub, Convert, One, Saturating, Zero,
 		},
-		AccountId32, DispatchErrorWithPostInfo, MultiSignature,
+		AccountId32, DispatchErrorWithPostInfo,
 	};
 	use sp_std::{fmt::Debug, vec::Vec};
 
@@ -443,17 +441,20 @@ pub mod pallet {
 				Proof::Ethereum(eth_proof) => {
 					let reward_account_encoded =
 						reward_account.using_encoded(|x| hex::encode(x).as_bytes().to_vec());
-					let eth_address =
-						Self::ethereum_recover(prefix, &reward_account_encoded, &eth_proof)
-							.ok_or(Error::<T>::InvalidProof)?;
+					let eth_address = signature_verification::ethereum_recover(
+						prefix,
+						&reward_account_encoded,
+						&eth_proof,
+					)
+					.ok_or(Error::<T>::InvalidProof)?;
 					Result::<_, DispatchError>::Ok(Identity::Ethereum(eth_address))
 				},
 				Proof::RelayChain(relay_account, relay_proof) => {
 					ensure!(
-						Self::verify_relay(
+						signature_verification::verify_relay(
 							prefix,
 							reward_account.clone(),
-							relay_account.clone(),
+							relay_account.clone().into(),
 							&relay_proof
 						),
 						Error::<T>::InvalidProof
@@ -463,7 +464,7 @@ pub mod pallet {
 				Proof::Cosmos(cosmos_address, cosmos_proof) => {
 					let reward_account_encoded =
 						reward_account.using_encoded(|x| hex::encode(x).as_bytes().to_vec());
-					let cosmos_address = Self::cosmos_recover(
+					let cosmos_address = signature_verification::cosmos_recover(
 						prefix,
 						&reward_account_encoded,
 						cosmos_address,
@@ -474,102 +475,6 @@ pub mod pallet {
 				},
 			}?;
 			Ok(identity)
-		}
-
-		/// Verify the proof is valid for a given relay account.
-		pub(crate) fn verify_relay(
-			prefix: &[u8],
-			reward_account: <T as frame_system::Config>::AccountId,
-			relay_account: T::RelayChainAccountId,
-			proof: &MultiSignature,
-		) -> bool {
-			// Polkadot.js wrapper tags
-			const WRAPPED_PREFIX: &[u8] = b"<Bytes>";
-			const WRAPPED_POSTFIX: &[u8] = b"</Bytes>";
-			let mut msg = WRAPPED_PREFIX.to_vec();
-
-			msg.append(&mut prefix.to_vec());
-			msg.append(&mut reward_account.using_encoded(|x| hex::encode(x).as_bytes().to_vec()));
-			msg.append(&mut WRAPPED_POSTFIX.to_vec());
-
-			proof.verify(&msg[..], &relay_account.into())
-		}
-
-		/// Recover the public key of an `eth_sign` signature.
-		///
-		/// Requires the original message.
-		pub(crate) fn ethereum_recover(
-			prefix: &[u8],
-			msg: &[u8],
-			EcdsaSignature(sig): &EcdsaSignature,
-		) -> Option<EthereumAddress> {
-			let msg = keccak_256(&Self::ethereum_signable_message(prefix, msg));
-			let mut address = EthereumAddress::default();
-
-			address.0.copy_from_slice(
-				&keccak_256(&sp_io::crypto::secp256k1_ecdsa_recover(sig, &msg).ok()?[..])[12..],
-			);
-
-			Some(address)
-		}
-
-		/// Sign a message to produce an `eth_sign` compatible signature.
-		pub(crate) fn ethereum_signable_message(prefix: &[u8], msg: &[u8]) -> Vec<u8> {
-			let mut length = prefix.len() + msg.len();
-			let mut msg_len = Vec::new();
-
-			while length > 0 {
-				msg_len.push(b'0' + (length % 10) as u8);
-				length /= 10;
-			}
-
-			let mut signed_message = b"\x19Ethereum Signed Message:\n".to_vec();
-			signed_message.extend(msg_len.into_iter().rev());
-			signed_message.extend_from_slice(prefix);
-			signed_message.extend_from_slice(msg);
-
-			signed_message
-		}
-
-		/// From a signature and message, will attempt to recover a Cosmos AccAddress.
-		///
-		/// Returns `None` if signature is invalid.
-		pub(crate) fn cosmos_recover(
-			prefix: &[u8],
-			msg: &[u8],
-			cosmos_address: CosmosAddress,
-			CosmosEcdsaSignature(sig): &CosmosEcdsaSignature,
-		) -> Option<CosmosAddress> {
-			let msg = sha2_256(&[prefix, msg].concat());
-
-			match cosmos_address {
-				CosmosAddress::Secp256k1(pub_key) => {
-					// Cosmos gives us a 64-byte, we convert it into the more standard 65-byte
-					// signature here
-					let sig: EcdsaSignature = CosmosEcdsaSignature(*sig).into();
-
-					if pub_key ==
-						sp_io::crypto::secp256k1_ecdsa_recover_compressed(&sig.0, &msg).ok()?
-					{
-						return Some(CosmosAddress::Secp256k1(pub_key))
-					}
-
-					None
-				},
-				CosmosAddress::Secp256r1(pub_key) => {
-					// Deconstruct `sig` into `r` and `s` values so we can construct a p256
-					// friendly signature
-					let mut r: [u8; 32] = [0; 32];
-					let mut s: [u8; 32] = [0; 32];
-					r.copy_from_slice(&sig[..32]);
-					s.copy_from_slice(&sig[32..64]);
-
-					let sig = Signature::from_scalars(r, s).ok()?;
-					let verify_key = VerifyingKey::from_sec1_bytes(&pub_key).ok()?;
-					let _ = verify_key.verify(&msg, &sig).ok()?;
-					Some(CosmosAddress::Secp256r1(pub_key))
-				},
-			}
 		}
 
 		/// Start an Airdrop at a given moment.
