+++ frame/composable-support/Cargo.toml
@@ -13,6 +13,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 frame-support = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
 
 sp-arithmetic = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
+sp-io = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
 sp-std = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.22" }
 
@@ -24,7 +25,9 @@ scale-info = { version = "2.1.1", default-features = false, features = [
 ] }
 
 is_sorted = { version = "0.1.1", default-features = false }
+hex = { version = "0.4", default-features = false, features = ["alloc"] }
 num-traits = { version = "0.2.14", default-features = false }
+p256 = { version = "0.10.1", default-features = false, features = ["ecdsa", "ecdsa-core"] }
 rustc-hex = { version = "2.1.0", default-features = false }
 serde = { version = "1.0.136", features = ["derive"], optional = true }
