+++ frame/composable-support/src/lib.rs
@@ -16,5 +16,6 @@ pub mod abstractions;
 pub mod collections;
 pub mod math;
 pub mod rpc_helpers;
+pub mod signature_verification;
 pub mod types;
 pub mod validation;
